requirements.txt
tensorflow==2.4.1
absl-py==0.12.0
numpy==1.19.5
imageio==2.9.0
matplotlib==3.3.4
scipy==1.6.3
pyyaml==5.4.1
h5py==2.10.0
